+++ src/TextSearch.cpp
@@ -9,7 +9,7 @@ TextSearch::TextSearch(const sserialize::UByteArrayAdapter& d, const sserialize:
 	SSERIALIZE_VERSION_MISSMATCH_CHECK(LIBOSCAR_TEXT_SEARCH_VERSION, d.at(0), "liboscar::TextSearch");
 	
 	sserialize::Static::Array<sserialize::UByteArrayAdapter> tmp(d+1);
-	for(uint32_t i = 0, s = tmp.size(); i < s; ++i) {
+	for(uint32_t i = 0, s = (uint32_t) tmp.size(); i < s; ++i) {
 		sserialize::UByteArrayAdapter td(tmp.at(i));
 		switch (td.at(0)) {
 		case ITEMS:
